refactor versions into seperate packages
