+++ jobs-api-server/src/main/java/org/zowe/jobs/v2/services/JobsServiceV2.java
@@ -7,7 +7,7 @@
  *
  * Copyright IBM Corporation 2018, 2019
  */
-package org.zowe.jobs.services;
+package org.zowe.jobs.v2.services;
 
 import org.zowe.api.common.exceptions.ZoweApiException;
 import org.zowe.api.common.model.ItemsWrapper;
@@ -16,7 +16,7 @@ import org.zowe.jobs.model.JobFile;
 import org.zowe.jobs.model.JobFileContent;
 import org.zowe.jobs.model.JobStatus;
 
-public interface JobsServiceV1 {
+public interface JobsServiceV2 {
 
     ItemsWrapper<Job> getJobs(String prefix, String owner, JobStatus status) throws ZoweApiException;
