+++ jobs-api-server/src/main/java/org/zowe/jobs/v2/services/zosmf/AbstractZosmfJobsRequestRunner.java
@@ -7,7 +7,7 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services.zosmf;
+package org.zowe.jobs.v2.services.zosmf;
 
 import com.google.gson.JsonObject;
