+++ jobs-api-server/src/main/java/org/zowe/jobs/v2/services/zosmf/GetJobFilesZosmfRequestRunner.java
@@ -7,34 +7,38 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services.zosmf;
+package org.zowe.jobs.v2.services.zosmf;
 
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.RequestBuilder;
 import org.zowe.api.common.connectors.zosmf.ZosmfConnectorV2;
 import org.zowe.api.common.exceptions.ZoweApiRestException;
+import org.zowe.api.common.model.ItemsWrapper;
 import org.zowe.api.common.utils.ResponseCache;
-import org.zowe.jobs.model.Job;
+import org.zowe.jobs.model.JobFile;
 
 import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
+import java.util.List;
 
-public class GetJobZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<Job> {
+public class GetJobFilesZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<ItemsWrapper<JobFile>> {
 
     private String jobName;
     private String jobId;
 
-    public GetJobZosmfRequestRunner(String jobName, String jobId) {
+    public GetJobFilesZosmfRequestRunner(String jobName, String jobId) {
         this.jobName = jobName;
         this.jobId = jobId;
     }
 
     @Override
     protected RequestBuilder prepareQuery(ZosmfConnectorV2 zosmfConnector) throws URISyntaxException {
-        String urlPath = String.format("restjobs/jobs/%s/%s", jobName, jobId); //$NON-NLS-1$
+        String urlPath = String.format("restjobs/jobs/%s/%s/files", jobName, jobId); //$NON-NLS-1$
         URI requestUrl = zosmfConnector.getFullUrl(urlPath);
         return RequestBuilder.get(requestUrl);
     }
@@ -45,12 +49,24 @@ public class GetJobZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<Job
     }
 
     @Override
-    protected Job getResult(ResponseCache responseCache) throws IOException {
-        return getJobFromJson(responseCache.getEntityAsJsonObject());
+    protected ItemsWrapper<JobFile> getResult(ResponseCache responseCache) throws IOException {
+        JsonElement jsonResponse = responseCache.getEntityAsJson();
+        List<JobFile> jobFiles = new ArrayList<>();
+        for (JsonElement jsonElement : jsonResponse.getAsJsonArray()) {
+            jobFiles.add(getJobFileFromJson(jsonElement.getAsJsonObject()));
+        }
+        return new ItemsWrapper<JobFile>(jobFiles);
     }
 
     @Override
     protected ZoweApiRestException createException(JsonObject jsonResponse, int statusCode) {
         return createJobNotFoundExceptions(jsonResponse, statusCode, jobName, jobId);
+    }
+
+    private static JobFile getJobFileFromJson(JsonObject returned) {
+        return JobFile.builder().id(returned.get("id").getAsLong()).ddName(returned.get("ddname").getAsString())
+            .recordFormat(returned.get("recfm").getAsString()).recordLength(returned.get("lrecl").getAsLong())
+            .byteCount(returned.get("byte-count").getAsLong()).recordCount(returned.get("record-count").getAsLong())
+            .build();
     }
 }
