+++ jobs-api-server/src/main/java/org/zowe/jobs/v2/services/zosmf/PurgeJobZosmfRequestRunner.java
@@ -7,13 +7,13 @@
  *
  * Copyright IBM Corporation 2019
  */
-package org.zowe.jobs.services.zosmf.v1;
+package org.zowe.jobs.v2.services.zosmf;
 
 import com.google.gson.JsonObject;
 
 import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.RequestBuilder;
-import org.zowe.api.common.connectors.zosmf.ZosmfConnectorV1;
+import org.zowe.api.common.connectors.zosmf.ZosmfConnectorV2;
 import org.zowe.api.common.exceptions.ZoweApiRestException;
 import org.zowe.api.common.utils.ResponseCache;
 
@@ -37,7 +37,7 @@ public class PurgeJobZosmfRequestRunner extends AbstractZosmfJobsRequestRunner<V
     }
 
     @Override
-    protected RequestBuilder prepareQuery(ZosmfConnectorV1 zosmfConnector) throws URISyntaxException {
+    protected RequestBuilder prepareQuery(ZosmfConnectorV2 zosmfConnector) throws URISyntaxException {
         String urlPath = String.format("restjobs/jobs/%s/%s", jobName, jobId); //$NON-NLS-1$
         URI requestUrl = zosmfConnector.getFullUrl(urlPath);
         return RequestBuilder.delete(requestUrl);
