+++ jobs-api-server/src/test/java/org/zowe/jobs/services/zosmf/GetJobFileContentZosmfRequestRunnerTest.java
@@ -17,6 +17,7 @@ import org.zowe.jobs.exceptions.JobFileIdNotFoundException;
 import org.zowe.jobs.exceptions.JobIdNotFoundException;
 import org.zowe.jobs.exceptions.JobNameNotFoundException;
 import org.zowe.jobs.model.JobFileContent;
+import org.zowe.jobs.v2.services.zosmf.GetJobFileContentZosmfRequestRunner;
 
 import java.io.IOException;
