+++ jobs-api-server/src/test/java/org/zowe/jobs/services/zosmf/PurgeJobZosmfRequestRunnerTest.java
@@ -13,6 +13,7 @@ import org.apache.http.HttpStatus;
 import org.apache.http.client.methods.RequestBuilder;
 import org.junit.Test;
 import org.zowe.jobs.exceptions.JobNameNotFoundException;
+import org.zowe.jobs.v2.services.zosmf.PurgeJobZosmfRequestRunner;
 
 import static org.mockito.Mockito.when;
