+++ jobs-api-server/src/test/java/org/zowe/jobs/services/zosmf/ZosmfJobsServiceTest.java
@@ -30,6 +30,15 @@ import org.zowe.jobs.model.Job;
 import org.zowe.jobs.model.JobFile;
 import org.zowe.jobs.model.JobFileContent;
 import org.zowe.jobs.model.JobStatus;
+import org.zowe.jobs.v2.services.zosmf.GetJobFileContentZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.GetJobFilesZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.GetJobZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.GetJobsZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.ModifyJobZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.PurgeJobZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.SubmitJobFileZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.SubmitJobStringZosmfRequestRunner;
+import org.zowe.jobs.v2.services.zosmf.ZosmfJobsServiceV2;
 
 import java.util.Arrays;
 import java.util.List;
@@ -40,17 +49,17 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({ ZosmfJobsService.class })
+@PrepareForTest({ ZosmfJobsServiceV2.class })
 public class ZosmfJobsServiceTest extends ZoweApiTest {
 
     @Mock
     ZosmfConnector zosmfConnector;
 
-    ZosmfJobsService jobsService;
+    ZosmfJobsServiceV2 jobsService;
 
     @Before
     public void setUp() throws Exception {
-        jobsService = new ZosmfJobsService();
+        jobsService = new ZosmfJobsServiceV2();
         jobsService.zosmfConnector = zosmfConnector;
     }
